requirements.txt
pandas
scikit-learn
numpy
joblib
fastapi
uvicorn
evidently
requests
PyYAML
ipykernel
matplotlib
seaborn
dagshub
mlflow
airflow
dvc